routes: migrate POST /api/goals + DELETE goals/:id + gates signal + sessions/continue (4 routes)
Move to:
- src/server/routes/goals.ts: POST /api/goals (create with auto-start-team)
  + DELETE /api/goals/:id (archive + remote-branch cleanup)
- src/server/routes/gates.ts: POST /api/goals/:id/gates/:gateId/signal
  (full verification harness flow with cascade reset)
- src/server/routes/sessions.ts: POST /api/sessions/:archivedId/continue
  (lossless transcript clone + resume)

Drop now-unused imports (detectPrimaryBranch, hasTransitiveDep, execGitSafe,
deleteRemoteGoalBranches). Underscore-prefix more handleApiRoute params.

Net: server.ts -511 lines (2300 → 1789). No behaviour change.
